requirements_streamlit.txt
# Streamlit App Requirements
# Complete dependencies for the NMLS Search Intelligence Platform

# Web Framework
streamlit>=1.28.0

# Data Processing & Visualization
pandas>=2.1.3
plotly>=5.17.0
numpy>=1.24.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Natural Language Processing (from your existing setup)
anthropic>=0.40.0
sentence-transformers>=2.2.2
torch>=2.0.0

# API Framework (for backend integration)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.0

# Data Validation
pydantic>=2.5.0

# HTML Processing (from your existing pipeline)
beautifulsoup4>=4.12.2
lxml>=4.9.0

# Environment Management
python-dotenv>=1.0.0

# Additional utilities
python-multipart>=0.0.6 